test(dropdown): teardown correctly
