+++ src/components/Header/utils.ts
@@ -1,0 +1,11 @@
+import { noOp } from '../../utils';
+
+export const defaultHeaderContext = {
+  defaultMobileMenuLabel: '',
+  expandedItem: '',
+  setExpandedItem: noOp,
+  isExpanded: false,
+  onSelect: noOp,
+  isSearchExpanded: false,
+  setIsSearchExpanded: noOp,
+};
